+++ src/Systems/DX9RenderSystem/DX9RenderSystem.h
@@ -107,7 +107,7 @@ namespace Mengine
         bool beginScene() override;
         void endScene() override;
         void swapBuffers() override;
-        void clearFrameBuffer( uint32_t _frameBufferTypes, const Color & _color, float _depth, uint32_t _stencil ) override;
+        void clearFrameBuffer( uint32_t _frameBufferTypes, const Color & _color, double _depth, int32_t _stencil ) override;
 
         void setScissor( const Viewport & _viewport ) override;
         void removeScissor() override;
